+++ packages/icons/src/components/AppSwitcher.tsx
@@ -15,7 +15,7 @@ export const AppSwitcherIcon = forwardRef<SVGSVGElement, AppSwitcherIconProps>(
         ref={ref}
         {...props}
       >
-        <path d="M3 0H0v3h3V0zm4.5 0h-3v3h3V0zM9 3V0h3v3H9zm0 1.5v3h3v-3H9zM9 12V9h3v3H9zM4.5 9v3h3V9h-3zM0 9h3v3H0V9zm3-4.5H0v3h3v-3zm1.5 0h3v3h-3v-3z" />
+        <path d="M3 0H0v3h3V0Zm4.5 0h-3v3h3V0ZM9 3V0h3v3H9Zm0 1.5v3h3v-3H9ZM9 12V9h3v3H9ZM4.5 9v3h3V9h-3ZM0 9h3v3H0V9Zm3-4.5H0v3h3v-3Zm1.5 0h3v3h-3v-3Z" />
       </Icon>
     );
   }