+++ packages/icons/src/components/ChevronLeft.tsx
@@ -15,7 +15,11 @@ export const ChevronLeftIcon = forwardRef<SVGSVGElement, ChevronLeftIconProps>(
         ref={ref}
         {...props}
       >
-        <path d="M4.692 6 9.5 10.401 8.508 11.5 2.5 6 8.508.5 9.5 1.599 4.692 6z" />
+        <path
+          fillRule="evenodd"
+          d="M4.692 6 9.5 10.401 8.508 11.5 2.5 6 8.508.5 9.5 1.599 4.692 6Z"
+          clipRule="evenodd"
+        />
       </Icon>
     );
   }