+++ packages/icons/src/components/ClockSolid.tsx
@@ -15,7 +15,11 @@ export const ClockSolidIcon = forwardRef<SVGSVGElement, ClockSolidIconProps>(
         ref={ref}
         {...props}
       >
-        <path d="M12 6A6 6 0 1 1 0 6a6 6 0 0 1 12 0zM6 3v3H3v1h4V3H6z" />
+        <path
+          fillRule="evenodd"
+          d="M12 6A6 6 0 1 1 0 6a6 6 0 0 1 12 0ZM6 3v3H3v1h4V3H6Z"
+          clipRule="evenodd"
+        />
       </Icon>
     );
   }