+++ packages/icons/src/components/ColumnChooser.tsx
@@ -17,7 +17,11 @@ export const ColumnChooserIcon = forwardRef<
       ref={ref}
       {...props}
     >
-      <path d="M0 0h12v12H0V0zm1 3v1h3v1H1v1h3v1H1v1h3v1H1v2h4V3H1zm5 0v1h3v1H6v1h3v1H6v1h3v1H6v2h5V3H6zm5-2H1v1h10V1z" />
+      <path
+        fillRule="evenodd"
+        d="M0 0h12v12H0V0Zm1 3v1h3v1H1v1h3v1H1v1h3v1H1v2h4V3H1Zm5 0v1h3v1H6v1h3v1H6v1h3v1H6v2h5V3H6Zm5-2H1v1h10V1Z"
+        clipRule="evenodd"
+      />
     </Icon>
   );
 });