+++ packages/icons/src/components/Save.tsx
@@ -15,8 +15,16 @@ export const SaveIcon = forwardRef<SVGSVGElement, SaveIconProps>(
         ref={ref}
         {...props}
       >
-        <path d="M10 6v4H2V6h8zM9 7H3v2h6V7z" />
-        <path d="M0 0h8.707L12 3.293V12H0V0zm3 1H1v10h10V3.707L8.293 1H8v3H3V1zm4 0H4v2h3V1z" />
+        <path
+          fillRule="evenodd"
+          d="M10 6H2v4h8V6ZM9 7H3v2h6V7Z"
+          clipRule="evenodd"
+        />
+        <path
+          fillRule="evenodd"
+          d="M0 0v12h12V3.293L8.707 0H0Zm3 1H1v10h10V3.707L8.293 1H8v3H3V1Zm4 0H4v2h3V1Z"
+          clipRule="evenodd"
+        />
       </Icon>
     );
   }